separates unit and integration tests; npm test still runs them all
