--- test/test.js
@@ -1,3 +1,0 @@
-'use strict';
-
-describe('example test', function() {});
