+++ test/unit/helpers.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var helpers = require('../src/helpers.js');
+var helpers = require('../../src/helpers.js');
 var assert = require('assert');
 
 describe('helpers', function () {
